+++ src/components/Footer.jsx
@@ -1,5 +1,5 @@
-import { socialLinks } from "./social_links";
-import { random_keyID } from "./random_keyID.js";
+import { socialLinks } from "./component_data/social_links";
+import { random_keyID } from "./component_data/random_keyID";
 
 const Footer = () => {
   return (
