+++ src/components/Projects.jsx
@@ -1,6 +1,6 @@
 "use client";
-import projectData from "./project_data.js";
-import { random_keyID } from "./random_keyID.js";
+import projectData from "./component_data/project_data";
+import { random_keyID } from "./component_data/random_keyID";
 import { useState } from "react";
 import Image from "next/image.js";
 import { forwardRef } from "react";
