+++ src/components/component_data/hobbies.js
