+++ src/components/component_data/social_links.js
